internal: move code to where it's used and reduce visibility
